// BE-Sgroup-2025-4-6-/src/models/inputTypeSchemas/CardTagWhereInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFilterSchema } from './StringFilterSchema';
import { DateTimeFilterSchema } from './DateTimeFilterSchema';
import { CardScalarRelationFilterSchema } from './CardScalarRelationFilterSchema';
import { CardWhereInputSchema } from './CardWhereInputSchema';

export const CardTagWhereInputSchema: z.ZodType<Prisma.CardTagWhereInput> = z.strictObject({
  AND: z.union([ z.lazy(() => CardTagWhereInputSchema), z.lazy(() => CardTagWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => CardTagWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => CardTagWhereInputSchema), z.lazy(() => CardTagWhereInputSchema).array() ]).optional(),
  id: z.union([ z.lazy(() => StringFilterSchema), z.string() ]).optional(),
  cardId: z.union([ z.lazy(() => StringFilterSchema), z.string() ]).optional(),
  name: z.union([ z.lazy(() => StringFilterSchema), z.string() ]).optional(),
  color: z.union([ z.lazy(() => StringFilterSchema), z.string() ]).optional(),
  createdAt: z.union([ z.lazy(() => DateTimeFilterSchema), z.coerce.date() ]).optional(),
  card: z.union([ z.lazy(() => CardScalarRelationFilterSchema), z.lazy(() => CardWhereInputSchema) ]).optional(),
});

export default CardTagWhereInputSchema;